requirements/base.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
psycopg==3.1.18
channels==4.0.0
channels-redis==4.2.0
Pillow==10.2.0
celery==5.3.6
redis==5.0.1
django-treebeard==4.7.1
django-guardian==2.4.0